requirements.txt

pdfplumber==0.11.7
pymongo==4.14.1
python-dotenv==1.1.1


